impl(017): Wave 8 Stage D2 WI-1 — boot relocation into ValkeyBackend
Move the five Valkey-specific deployment-initialisation steps from
`Server::start_with_metrics` into `ValkeyBackend::initialize_deployment`
per RFC-017 §4 row 12:

  1. verify_valkey_version (reject pre-7.2)
  2. validate_or_create_partition_config
  3. initialize_waitpoint_hmac_secret
  4. ensure_library (FUNCTION LOAD)
  5. Lanes SADD seed

Ordering contract (load-bearing) documented on
`ValkeyBackend::connect_with_metrics` trait doc-comment and on
`initialize_deployment`. Order matches the pre-relocation body
byte-for-byte; no logic changes.

- New `crates/ff-backend-valkey/src/boot.rs` owns the step bodies,
  returning `EngineError` (version-too-low / partition-mismatch /
  redis-rejected map to `ValidationKind::Corruption`; transport faults
  use `backend_context(transport_fk(e), …)`).
- `ValkeyBackend::initialize_deployment(&self, secret, lanes, skip)`
  is the new entry point. `Server::start_with_metrics` calls it once,
  after dial + before engine/backend wiring.
- 478 lines of inline boot helpers + 245 lines of unit tests deleted
  from `server.rs`; 14 boot-module unit tests re-added under
  `ff-backend-valkey` with the same coverage (parse_valkey_version +
  extract_info_bodies + cluster-min fixture).
