+++ crates/ff-server/src/server.rs
@@ -24,8 +24,7 @@ use ff_core::keys::{
     IndexKeys, QuotaKeyContext,
 };
 use ff_core::partition::{
-    budget_partition, execution_partition, flow_partition, quota_partition, Partition,
-    PartitionConfig, PartitionFamily,
+    budget_partition, execution_partition, flow_partition, quota_partition, PartitionConfig,
 };
 use ff_core::state::{PublicState, StateVector};
 use ff_core::types::*;
@@ -446,63 +445,26 @@ impl Server {
         }
         tracing::info!("Valkey connection established");
 
-        // Step 1b: Verify Valkey version meets the RFC-011 §13 minimum (8.0).
-        // Tolerates a rolling upgrade via a 60s exponential-backoff budget
-        // per RFC-011 §9.17 — transient INFO errors during a node restart
-        // don't trip the check until the whole budget is exhausted.
-        verify_valkey_version(&client).await?;
-
-        // Step 2: Validate or create partition config
-        validate_or_create_partition_config(&client, &config.partition_config).await?;
-
-        // Step 2b: Install waitpoint HMAC secret into every execution partition
-        // (RFC-004 §Waitpoint Security). Fail-fast: if any partition fails,
-        // the server refuses to start — a partial install would silently
-        // reject signal deliveries on half the partitions.
-        initialize_waitpoint_hmac_secret(
-            &client,
-            &config.partition_config,
-            &config.waitpoint_hmac_secret,
-        )
-        .await?;
-
-        // Step 3: Load Lua library (skippable for tests where fixture already loaded)
-        if !config.skip_library_load {
-            tracing::info!("loading flowfabric Lua library");
-            ff_script::loader::ensure_library(&client)
-                .await
-                .map_err(ServerError::LibraryLoad)?;
-        } else {
-            tracing::info!("skipping library load (skip_library_load=true)");
-        }
-
-        // Step 3b: Seed the global lanes registry (`ff:idx:lanes`) with
-        // every lane from ServerConfig.lanes (issue #203).
-        //
-        // `EngineBackend::list_lanes` reads SMEMBERS from this global SET,
-        // so without a boot-time seed a freshly-provisioned deployment
-        // returns an empty lane list until the first execution is created
-        // on each lane. The Lua `ff_create_execution` path also SADDs on
-        // first-sight, which covers dynamic / post-boot lanes; this block
-        // covers the pre-declared set. SADD is idempotent.
-        //
-        // NOTE: `ff:idx:lanes` is intentionally NOT hash-tagged per
-        // partition — it is the single cross-partition global SET in the
-        // system (see `ff_core::keys::lanes_index_key`).
-        if !config.lanes.is_empty() {
-            let lane_strs: Vec<&str> = config.lanes.iter().map(|l| l.as_str()).collect();
-            let _: i64 = client
-                .cmd("SADD")
-                .arg(ff_core::keys::lanes_index_key().as_str())
-                .arg(lane_strs.as_slice())
-                .execute()
-                .await
-                .map_err(|e| crate::server::backend_context(e, "SADD ff:idx:lanes (seed)"))?;
-            tracing::info!(
-                lanes = ?config.lanes.iter().map(|l| l.as_str()).collect::<Vec<_>>(),
-                "seeded lanes index (ff:idx:lanes)"
-            );
-        }
+        // RFC-017 Wave 8 Stage D (§4 row 12): the five Valkey-specific
+        // deployment-initialisation steps (version verify, partition
+        // config, HMAC install, FUNCTION LOAD, lanes seed) now live
+        // behind `ValkeyBackend::initialize_deployment`. Load-bearing
+        // ordering contract preserved in that method's doc-comment.
+        // The call is sequenced after the connection but before the
+        // engine + backend wiring below so the pre-relocation boot
+        // ordering is byte-for-byte identical.
+        let init_backend = ff_backend_valkey::ValkeyBackend::from_client_and_partitions(
+            client.clone(),
+            config.partition_config,
+        );
+        init_backend
+            .initialize_deployment(
+                &config.waitpoint_hmac_secret,
+                &config.lanes,
+                config.skip_library_load,
+            )
+            .await
+            .map_err(|e| ServerError::Engine(Box::new(e)))?;
 
         // Step 4: Start engine with scanners
         // Build a fresh EngineConfig rather than cloning (EngineConfig doesn't derive Clone).
@@ -2396,485 +2358,6 @@ fn wire_str_to_stream_cursor(s: &str) -> ff_core::contracts::StreamCursor {
     }
 }
 
-// ── Valkey version check (RFC-011 §13) ──
-
-/// Minimum Valkey version the engine requires (see RFC-011 §13). 7.2 is the
-/// release where Valkey Functions and RESP3 stabilized — the primitives the
-/// co-location design and typed FCALL wrappers actually depend on.
-const REQUIRED_VALKEY_MAJOR: u32 = 7;
-const REQUIRED_VALKEY_MINOR: u32 = 2;
-
-/// Upper bound on the rolling-upgrade retry window (RFC-011 §9.17). A Valkey
-/// node cycling through SIGTERM → restart typically completes in well under
-/// 60s; this budget is generous without letting a truly-stuck cluster hang
-/// boot indefinitely.
-const VERSION_CHECK_RETRY_BUDGET: Duration = Duration::from_secs(60);
-
-/// Verify the connected Valkey reports a version ≥ 7.2.
-///
-/// Per RFC-011 §9.17, during a rolling upgrade the node we happen to connect
-/// to may temporarily be pre-upgrade while others are post-upgrade. The check
-/// tolerates this by retrying the whole verification (including low-version
-/// responses) with exponential backoff, capped at a 60s budget.
-///
-/// **Retries on:**
-/// - Low-version responses (below `(REQUIRED_VALKEY_MAJOR, REQUIRED_VALKEY_MINOR)`)
-///   — may resolve as the rolling upgrade progresses onto the connected node.
-/// - Retryable ferriskey transport errors — connection refused,
-///   `BusyLoadingError`, `ClusterDown`, etc., classified via
-///   `ff_script::retry::is_retryable_kind`.
-/// - Missing/unparsable version field — treated as transient (fresh-boot
-///   server may not have the INFO fields populated yet). Reads
-///   `valkey_version` when present (authoritative on Valkey 8.0+), falls
-///   back to `redis_version` for Valkey 7.x.
-///
-/// **Does NOT retry on:**
-/// - Non-retryable transport errors (auth failures, permission denied,
-///   invalid client config) — these are operator misconfiguration, not
-///   transient cluster state; fast-fail preserves a clear signal.
-///
-/// On budget exhaustion, returns the last observed error.
-async fn verify_valkey_version(client: &Client) -> Result<(), ServerError> {
-    let deadline = tokio::time::Instant::now() + VERSION_CHECK_RETRY_BUDGET;
-    let mut backoff = Duration::from_millis(200);
-    loop {
-        let (should_retry, err_for_budget_exhaust, log_detail): (bool, ServerError, String) =
-            match query_valkey_version(client).await {
-                Ok((detected_major, detected_minor))
-                    if (detected_major, detected_minor)
-                        >= (REQUIRED_VALKEY_MAJOR, REQUIRED_VALKEY_MINOR) =>
-                {
-                    tracing::info!(
-                        detected_major,
-                        detected_minor,
-                        required_major = REQUIRED_VALKEY_MAJOR,
-                        required_minor = REQUIRED_VALKEY_MINOR,
-                        "Valkey version accepted"
-                    );
-                    return Ok(());
-                }
-                Ok((detected_major, detected_minor)) => (
-                    // Low version — may be a rolling-upgrade stale node.
-                    // Retry within budget; after exhaustion, the cluster
-                    // is misconfigured and fast-fail is the correct signal.
-                    true,
-                    ServerError::ValkeyVersionTooLow {
-                        detected: format!("{detected_major}.{detected_minor}"),
-                        required: format!("{REQUIRED_VALKEY_MAJOR}.{REQUIRED_VALKEY_MINOR}"),
-                    },
-                    format!(
-                        "detected={detected_major}.{detected_minor} < required={REQUIRED_VALKEY_MAJOR}.{REQUIRED_VALKEY_MINOR}"
-                    ),
-                ),
-                Err(e) => {
-                    // Only retry if the underlying Valkey error is retryable
-                    // by kind. Auth / permission / invalid-config should fast-
-                    // fail so operators see the true root cause immediately,
-                    // not a 60s hang followed by a generic "transient" error.
-                    let retryable = e
-                        .backend_kind()
-                        .map(|k| k.is_retryable())
-                        // Non-backend errors (parse, missing field, operation
-                        // failures) are treated as transient — a fresh-boot
-                        // Valkey may not have redis_version populated yet.
-                        .unwrap_or(true);
-                    let detail = e.to_string();
-                    (retryable, e, detail)
-                }
-            };
-
-        if !should_retry {
-            return Err(err_for_budget_exhaust);
-        }
-        if tokio::time::Instant::now() >= deadline {
-            return Err(err_for_budget_exhaust);
-        }
-        tracing::warn!(
-            backoff_ms = backoff.as_millis() as u64,
-            detail = %log_detail,
-            "valkey version check transient failure; retrying"
-        );
-        tokio::time::sleep(backoff).await;
-        backoff = (backoff * 2).min(Duration::from_secs(5));
-    }
-}
-
-/// Run `INFO server` and extract the `(major, minor)` components of the
-/// Valkey version.
-///
-/// Returns `Err` on transport errors, missing field, or unparsable version.
-/// Handles three response shapes:
-///
-/// - **Standalone:** single string body; parse directly.
-/// - **Cluster (RESP3 map):** `INFO` returns a map keyed by node address;
-///   every node runs the same version in a healthy deployment, so we pick
-///   one entry and parse it. Divergent versions during a rolling upgrade
-///   are handled by the outer retry loop.
-/// - **Empty / unexpected:** surfaces as `OperationFailed` with context.
-async fn query_valkey_version(client: &Client) -> Result<(u32, u32), ServerError> {
-    let raw: Value = client
-        .cmd("INFO")
-        .arg("server")
-        .execute()
-        .await
-        .map_err(|e| crate::server::backend_context(e, "INFO server"))?;
-    let bodies = extract_info_bodies(&raw)?;
-    // Cluster: return the minimum (major, minor) across all nodes so a stale
-    // pre-upgrade replica cannot hide behind an already-upgraded primary.
-    // Standalone: exactly one body. The outer retry loop tolerates rolling
-    // upgrades — a briefly-low minimum gets retried; a persistently-low one
-    // exits with the structured floor error.
-    let mut min_version: Option<(u32, u32)> = None;
-    for body in &bodies {
-        let version = parse_valkey_version(body)?;
-        min_version = Some(match min_version {
-            None => version,
-            Some(existing) => existing.min(version),
-        });
-    }
-    min_version.ok_or_else(|| {
-        ServerError::OperationFailed(
-            "valkey version check: cluster INFO returned no node bodies".into(),
-        )
-    })
-}
-
-/// Normalize an `INFO server` response to one string body per node.
-///
-/// Standalone returns a single body. Cluster (RESP3 map keyed by node address)
-/// returns every node's body — the caller must consider all of them to reject
-/// a mixed-version cluster where one stale node is below the floor.
-fn extract_info_bodies(raw: &Value) -> Result<Vec<String>, ServerError> {
-    match raw {
-        Value::BulkString(bytes) => Ok(vec![String::from_utf8_lossy(bytes).into_owned()]),
-        Value::VerbatimString { text, .. } => Ok(vec![text.clone()]),
-        Value::SimpleString(s) => Ok(vec![s.clone()]),
-        Value::Map(entries) => {
-            if entries.is_empty() {
-                return Err(ServerError::OperationFailed(
-                    "valkey version check: cluster INFO returned empty map".into(),
-                ));
-            }
-            let mut out = Vec::with_capacity(entries.len());
-            for (_, body) in entries {
-                out.extend(extract_info_bodies(body)?);
-            }
-            Ok(out)
-        }
-        other => Err(ServerError::OperationFailed(format!(
-            "valkey version check: unexpected INFO shape: {other:?}"
-        ))),
-    }
-}
-
-/// Extract the `(major, minor)` components of the Valkey version from an
-/// `INFO server` response body. Pure parser — pulled out of
-/// [`query_valkey_version`] so it is unit-testable without a live Valkey.
-///
-/// **Prefers `valkey_version:`** (introduced in Valkey 8.0+; this is the real
-/// server version on 8.x/9.x, which pin `redis_version:7.2.4` for
-/// Redis-client compatibility).
-///
-/// **Falls back to `redis_version:` only when the body carries an affirmative
-/// `server_name:valkey` field.** `server_name:` was introduced in Valkey 7.2,
-/// which is our floor — so every floor-compliant Valkey deployment carries
-/// the marker. Redis does not emit `server_name:valkey`, which is how we
-/// reject a Redis backend that would otherwise look identical to Valkey 7.x
-/// at the `redis_version:` level.
-fn parse_valkey_version(info: &str) -> Result<(u32, u32), ServerError> {
-    let extract_major_minor = |line: &str| -> Result<(u32, u32), ServerError> {
-        let trimmed = line.trim();
-        let mut parts = trimmed.split('.');
-        let major_str = parts.next().unwrap_or("").trim();
-        if major_str.is_empty() {
-            return Err(ServerError::OperationFailed(format!(
-                "valkey version check: empty version field in '{trimmed}'"
-            )));
-        }
-        let major = major_str.parse::<u32>().map_err(|_| {
-            ServerError::OperationFailed(format!(
-                "valkey version check: non-numeric major in '{trimmed}'"
-            ))
-        })?;
-        // Minor is required — a bare major ("7") cannot be compared against
-        // the (major, minor) floor reliably. Valkey always reports
-        // major.minor.patch for INFO, so missing minor is a real parse error.
-        let minor_str = parts.next().unwrap_or("").trim();
-        if minor_str.is_empty() {
-            return Err(ServerError::OperationFailed(format!(
-                "valkey version check: missing minor component in '{trimmed}'"
-            )));
-        }
-        let minor = minor_str.parse::<u32>().map_err(|_| {
-            ServerError::OperationFailed(format!(
-                "valkey version check: non-numeric minor in '{trimmed}'"
-            ))
-        })?;
-        Ok((major, minor))
-    };
-    // Prefer valkey_version (authoritative on Valkey 8.0+).
-    if let Some(valkey_line) = info
-        .lines()
-        .find_map(|line| line.strip_prefix("valkey_version:"))
-    {
-        return extract_major_minor(valkey_line);
-    }
-    // No valkey_version — could be Valkey 7.2 (which doesn't emit the field)
-    // or could be Redis. Require an affirmative server_name:valkey marker
-    // before falling back to redis_version. This rejects Redis backends,
-    // which don't emit server_name:valkey.
-    let server_is_valkey = info
-        .lines()
-        .map(str::trim)
-        .any(|line| line.eq_ignore_ascii_case("server_name:valkey"));
-    if !server_is_valkey {
-        return Err(ServerError::OperationFailed(
-            "valkey version check: INFO missing valkey_version and server_name:valkey marker \
-             (unsupported backend — FlowFabric requires Valkey >= 7.2; Redis is not supported)"
-                .into(),
-        ));
-    }
-    // Valkey 7.x fallback. 8.x+ pins redis_version:7.2.4 for Redis-client
-    // compat, so reading it there would under-report — but 8.x+ is handled
-    // by the valkey_version branch above, so we only reach here on 7.x.
-    if let Some(redis_line) = info
-        .lines()
-        .find_map(|line| line.strip_prefix("redis_version:"))
-    {
-        return extract_major_minor(redis_line);
-    }
-    Err(ServerError::OperationFailed(
-        "valkey version check: INFO has server_name:valkey but no redis_version or valkey_version field"
-            .into(),
-    ))
-}
-
-// ── Partition config validation ──
-
-/// Validate or create the `ff:config:partitions` key on first boot.
-///
-/// If the key exists, its values must match the server's config.
-/// If it doesn't exist, create it (first boot).
-async fn validate_or_create_partition_config(
-    client: &Client,
-    config: &PartitionConfig,
-) -> Result<(), ServerError> {
-    let key = keys::global_config_partitions();
-
-    let existing: HashMap<String, String> = client
-        .hgetall(&key)
-        .await
-        .map_err(|e| crate::server::backend_context(e, format!("HGETALL {key}")))?;
-
-    if existing.is_empty() {
-        // First boot — create the config
-        tracing::info!("first boot: creating {key}");
-        client
-            .hset(&key, "num_flow_partitions", &config.num_flow_partitions.to_string())
-            .await
-            .map_err(|e| crate::server::backend_context(e, "HSET num_flow_partitions"))?;
-        client
-            .hset(&key, "num_budget_partitions", &config.num_budget_partitions.to_string())
-            .await
-            .map_err(|e| crate::server::backend_context(e, "HSET num_budget_partitions"))?;
-        client
-            .hset(&key, "num_quota_partitions", &config.num_quota_partitions.to_string())
-            .await
-            .map_err(|e| crate::server::backend_context(e, "HSET num_quota_partitions"))?;
-        return Ok(());
-    }
-
-    // Validate existing config matches
-    let check = |field: &str, expected: u16| -> Result<(), ServerError> {
-        let stored: u16 = existing
-            .get(field)
-            .and_then(|v| v.parse().ok())
-            .unwrap_or(0);
-        if stored != expected {
-            return Err(ServerError::PartitionMismatch(format!(
-                "{field}: stored={stored}, config={expected}. \
-                 Partition counts are fixed at deployment time. \
-                 Either fix your config or migrate the data."
-            )));
-        }
-        Ok(())
-    };
-
-    check("num_flow_partitions", config.num_flow_partitions)?;
-    check("num_budget_partitions", config.num_budget_partitions)?;
-    check("num_quota_partitions", config.num_quota_partitions)?;
-
-    tracing::info!("partition config validated against stored {key}");
-    Ok(())
-}
-
-// ── Waitpoint HMAC secret bootstrap (RFC-004 §Waitpoint Security) ──
-
-/// Stable initial kid written on first boot. Rotation promotes to k2, k3, ...
-/// The kid is stored alongside the secret in every partition's hash so each
-/// FCALL can self-identify which secret produced a given token.
-const WAITPOINT_HMAC_INITIAL_KID: &str = "k1";
-
-/// Per-partition outcome of the HMAC bootstrap step. Collected across the
-/// parallel scan so we can emit aggregated logs once at the end.
-enum PartitionBootOutcome {
-    /// Partition already had a matching (kid, secret) pair.
-    Match,
-    /// Stored secret diverges from env — likely operator rotation; kept.
-    Mismatch,
-    /// Torn write (current_kid present, secret:<kid> missing); repaired.
-    Repaired,
-    /// Fresh partition; atomically installed env secret under kid=k1.
-    Installed,
-}
-
-/// Bounded in-flight concurrency for the startup fan-out. Large enough to
-/// turn a 256-partition install from ~15s sequential into ~1s on cross-AZ
-/// Valkey, small enough to leave a cold cluster breathing room for other
-/// Server::start work (library load, engine scanner spawn).
-const BOOT_INIT_CONCURRENCY: usize = 16;
-
-async fn init_one_partition(
-    client: &Client,
-    partition: Partition,
-    secret_hex: &str,
-) -> Result<PartitionBootOutcome, ServerError> {
-    let key = ff_core::keys::IndexKeys::new(&partition).waitpoint_hmac_secrets();
-
-    // Probe for an existing install. Fast path (fresh partition): HGET
-    // returns nil and we fall through to the atomic install below. Slow
-    // path: secret:<stored_kid> is then HGET'd once we know the kid name.
-    // (A previous version used a 2-field HMGET that included a fake
-    // `secret:probe` placeholder — vestigial from an abandoned
-    // optimization attempt, and confusing to read. Collapsed back to a
-    // single-field HGET.)
-    let stored_kid: Option<String> = client
-        .cmd("HGET")
-        .arg(&key)
-        .arg("current_kid")
-        .execute()
-        .await
-        .map_err(|e| crate::server::backend_context(e, format!("HGET {key} current_kid (init probe)")))?;
-
-    if let Some(stored_kid) = stored_kid {
-        // We didn't know the stored kid up front, so now HGET the real
-        // secret:<stored_kid> field. Two round-trips in the slow path; the
-        // fast path (fresh partition) stays at one.
-        let field = format!("secret:{stored_kid}");
-        let stored_secret: Option<String> = client
-            .hget(&key, &field)
-            .await
-            .map_err(|e| crate::server::backend_context(e, format!("HGET {key} secret:<kid> (init check)")))?;
-        if stored_secret.is_none() {
-            // Torn write from a previous boot: current_kid present but
-            // secret:<kid> missing. Without repair, mint returns
-            // "hmac_secret_not_initialized" on that partition forever.
-            // Repair in place with env secret. Not rotation — rotation
-            // always writes the secret first.
-            client
-                .hset(&key, &field, secret_hex)
-                .await
-                .map_err(|e| crate::server::backend_context(e, format!("HSET {key} secret:<kid> (repair torn write)")))?;
-            return Ok(PartitionBootOutcome::Repaired);
-        }
-        if stored_secret.as_deref() != Some(secret_hex) {
-            return Ok(PartitionBootOutcome::Mismatch);
-        }
-        return Ok(PartitionBootOutcome::Match);
-    }
-
-    // Fresh partition — install current_kid + secret:<kid> atomically in
-    // one HSET. Multi-field HSET is single-command atomic, so a crash
-    // can't leave current_kid without its secret.
-    let secret_field = format!("secret:{WAITPOINT_HMAC_INITIAL_KID}");
-    let _: i64 = client
-        .cmd("HSET")
-        .arg(&key)
-        .arg("current_kid")
-        .arg(WAITPOINT_HMAC_INITIAL_KID)
-        .arg(&secret_field)
-        .arg(secret_hex)
-        .execute()
-        .await
-        .map_err(|e| crate::server::backend_context(e, format!("HSET {key} (init waitpoint HMAC atomic)")))?;
-    Ok(PartitionBootOutcome::Installed)
-}
-
-/// Install the waitpoint HMAC secret on every execution partition.
-///
-/// Parallelized fan-out with bounded in-flight concurrency
-/// (`BOOT_INIT_CONCURRENCY`) so 256-partition boots finish in ~1s instead
-/// of ~15s sequential — the prior sequential loop was tight on K8s
-/// `initialDelaySeconds=30` defaults, especially cross-AZ. Fail-fast:
-/// the first per-partition error aborts boot.
-///
-/// Outcomes aggregate into mismatch/repaired counts (logged once at end)
-/// so operators see a single loud warning per fault class instead of 256
-/// per-partition lines.
-async fn initialize_waitpoint_hmac_secret(
-    client: &Client,
-    partition_config: &PartitionConfig,
-    secret_hex: &str,
-) -> Result<(), ServerError> {
-    use futures::stream::{FuturesUnordered, StreamExt};
-
-    let n = partition_config.num_flow_partitions;
-    tracing::info!(
-        partitions = n,
-        concurrency = BOOT_INIT_CONCURRENCY,
-        "installing waitpoint HMAC secret across {n} execution partitions"
-    );
-
-    let mut mismatch_count: u16 = 0;
-    let mut repaired_count: u16 = 0;
-    let mut pending: FuturesUnordered<_> = FuturesUnordered::new();
-    let mut next_index: u16 = 0;
-
-    loop {
-        while pending.len() < BOOT_INIT_CONCURRENCY && next_index < n {
-            let partition = Partition {
-                family: PartitionFamily::Execution,
-                index: next_index,
-            };
-            let client = client.clone();
-            let secret_hex = secret_hex.to_owned();
-            pending.push(async move {
-                init_one_partition(&client, partition, &secret_hex).await
-            });
-            next_index += 1;
-        }
-        match pending.next().await {
-            Some(res) => match res? {
-                PartitionBootOutcome::Match | PartitionBootOutcome::Installed => {}
-                PartitionBootOutcome::Mismatch => mismatch_count += 1,
-                PartitionBootOutcome::Repaired => repaired_count += 1,
-            },
-            None => break,
-        }
-    }
-
-    if repaired_count > 0 {
-        tracing::warn!(
-            repaired_partitions = repaired_count,
-            total_partitions = n,
-            "repaired {repaired_count} partitions with torn waitpoint HMAC writes \
-             (current_kid present but secret:<kid> missing, likely crash during prior boot)"
-        );
-    }
-
-    if mismatch_count > 0 {
-        tracing::warn!(
-            mismatched_partitions = mismatch_count,
-            total_partitions = n,
-            "stored/env secret mismatch on {mismatch_count} partitions — \
-             env FF_WAITPOINT_HMAC_SECRET ignored in favor of stored values; \
-             run POST /v1/admin/rotate-waitpoint-secret to sync"
-        );
-    }
-
-    tracing::info!(partitions = n, "waitpoint HMAC secret install complete");
-    Ok(())
-}
 
 /// Result of a waitpoint HMAC secret rotation across all execution partitions.
 #[derive(Debug, Clone, serde::Serialize)]
@@ -3986,251 +3469,6 @@ mod tests {
         assert_eq!(err.backend_kind(), None);
     }
 
-    // ── Valkey version check (RFC-011 §13) ──
-
-    #[test]
-    fn parse_valkey_version_prefers_valkey_version_over_redis_version() {
-        // Valkey 8.x+ pins redis_version to 7.2.4 for Redis-client compat
-        // and exposes the real version in valkey_version. Parser must use
-        // valkey_version when both are present.
-        let info = "\
-# Server\r\n\
-redis_version:7.2.4\r\n\
-valkey_version:9.0.3\r\n\
-server_mode:cluster\r\n\
-os:Linux\r\n";
-        assert_eq!(parse_valkey_version(info).unwrap(), (9, 0));
-    }
-
-    #[test]
-    fn parse_valkey_version_real_valkey_8_cluster_body() {
-        // Actual INFO server response observed on valkey/valkey:latest in
-        // cluster mode (CI matrix): redis_version compat-pinned to 7.2.4,
-        // valkey_version authoritative at 9.0.3.
-        let info = "\
-# Server\r\n\
-redis_version:7.2.4\r\n\
-server_name:valkey\r\n\
-valkey_version:9.0.3\r\n\
-valkey_release_stage:ga\r\n\
-redis_git_sha1:00000000\r\n\
-server_mode:cluster\r\n";
-        assert_eq!(parse_valkey_version(info).unwrap(), (9, 0));
-    }
-
-    #[test]
-    fn parse_valkey_version_falls_back_to_redis_version_on_valkey_7() {
-        // Valkey 7.x doesn't emit valkey_version, but does emit
-        // server_name:valkey; parser falls back to redis_version.
-        let info = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nfoo:bar\r\n";
-        assert_eq!(parse_valkey_version(info).unwrap(), (7, 2));
-    }
-
-    #[test]
-    fn parse_valkey_version_rejects_redis_backend() {
-        // Real Redis emits redis_version: but not server_name:valkey and
-        // not valkey_version:. Parser must reject this affirmatively so an
-        // operator pointing ff-server at a Redis instance fails loud instead
-        // of silently running against an unsupported backend.
-        let info = "\
-# Server\r\n\
-redis_version:7.4.0\r\n\
-redis_mode:standalone\r\n\
-os:Linux\r\n";
-        let err = parse_valkey_version(info).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        let msg = err.to_string();
-        assert!(
-            msg.contains("Redis is not supported") && msg.contains("server_name:valkey"),
-            "expected Redis-rejection message, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn parse_valkey_version_accepts_valkey_7_marker_case_insensitively() {
-        // INFO values are conventionally lowercase but be defensive.
-        let info = "redis_version:7.2.0\r\nSERVER_NAME:Valkey\r\n";
-        assert_eq!(parse_valkey_version(info).unwrap(), (7, 2));
-    }
-
-    #[test]
-    fn parse_valkey_version_errors_when_no_version_field() {
-        let info = "# Server\r\nfoo:bar\r\n";
-        let err = parse_valkey_version(info).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        assert!(
-            err.to_string().contains("missing"),
-            "expected 'missing' in message, got: {err}"
-        );
-    }
-
-    #[test]
-    fn parse_valkey_version_errors_on_non_numeric_major() {
-        let info = "valkey_version:invalid.x.y\n";
-        let err = parse_valkey_version(info).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        assert!(err.to_string().contains("non-numeric major"));
-    }
-
-    #[test]
-    fn parse_valkey_version_errors_on_non_numeric_minor() {
-        let info = "valkey_version:7.x.0\n";
-        let err = parse_valkey_version(info).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        assert!(err.to_string().contains("non-numeric minor"));
-    }
-
-    #[test]
-    fn parse_valkey_version_errors_on_missing_minor() {
-        // Bare major (no dot) — cannot be compared against a (major, minor)
-        // floor. Flag as a real parse error.
-        let info = "valkey_version:7\n";
-        let err = parse_valkey_version(info).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        assert!(err.to_string().contains("missing minor"));
-    }
-
-    #[test]
-    fn extract_info_bodies_unwraps_cluster_map_all_entries() {
-        // Simulates cluster-mode INFO response: map of node_addr → body.
-        // extract_info_bodies must return EVERY node's body so a stale
-        // pre-upgrade replica cannot hide behind an upgraded primary.
-        let body_a = "# Server\r\nredis_version:7.2.4\r\nvalkey_version:9.0.3\r\n";
-        let body_b = "# Server\r\nredis_version:7.2.4\r\nvalkey_version:8.0.0\r\n";
-        let map = Value::Map(vec![
-            (
-                Value::SimpleString("127.0.0.1:7000".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_a.to_string(),
-                },
-            ),
-            (
-                Value::SimpleString("127.0.0.1:7001".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_b.to_string(),
-                },
-            ),
-        ]);
-        let bodies = extract_info_bodies(&map).unwrap();
-        assert_eq!(bodies.len(), 2);
-        assert_eq!(bodies[0], body_a);
-        assert_eq!(bodies[1], body_b);
-    }
-
-    #[test]
-    fn extract_info_bodies_handles_simple_string() {
-        let body_text = "redis_version:7.2.4\r\nvalkey_version:9.0.3\r\n";
-        let v = Value::SimpleString(body_text.to_string());
-        let bodies = extract_info_bodies(&v).unwrap();
-        assert_eq!(bodies, vec![body_text.to_string()]);
-    }
-
-    #[test]
-    fn extract_info_bodies_rejects_empty_cluster_map() {
-        let map = Value::Map(vec![]);
-        let err = extract_info_bodies(&map).unwrap_err();
-        assert!(matches!(err, ServerError::OperationFailed(_)));
-        assert!(err.to_string().contains("empty map"));
-    }
-
-    /// End-to-end composition test for the cluster-min fix (issue #84):
-    /// `extract_info_bodies` → `parse_valkey_version` per node → min-reduce →
-    /// floor comparison. A mixed-version cluster where one node is 7.1.0 must
-    /// fail the gate, even if another node is already on 8.0.0 and that
-    /// node's entry appears first in the map.
-    #[test]
-    fn parse_valkey_version_min_across_cluster_map_picks_lowest() {
-        // node1 appears first and is above the floor. Pre-fix behavior
-        // (first-entry only) would accept. The min across all three nodes is
-        // (7, 1), below the (7, 2) floor, so the gate must reject.
-        let body_node1 = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nvalkey_version:8.0.0\r\n";
-        let body_node2 = "# Server\r\nredis_version:7.1.0\r\nserver_name:valkey\r\n";
-        let body_node3 = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nvalkey_version:7.2.0\r\n";
-        let map = Value::Map(vec![
-            (
-                Value::SimpleString("node1:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node1.to_string(),
-                },
-            ),
-            (
-                Value::SimpleString("node2:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node2.to_string(),
-                },
-            ),
-            (
-                Value::SimpleString("node3:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node3.to_string(),
-                },
-            ),
-        ]);
-
-        let bodies = extract_info_bodies(&map).unwrap();
-        let min = bodies
-            .iter()
-            .map(|b| parse_valkey_version(b).unwrap())
-            .min()
-            .unwrap();
-
-        assert_eq!(min, (7, 1), "min across cluster must be the lowest node");
-        assert!(
-            min < (REQUIRED_VALKEY_MAJOR, REQUIRED_VALKEY_MINOR),
-            "mixed-version cluster with 7.1.0 node must fail the (7,2) gate"
-        );
-    }
-
-    /// Companion to `parse_valkey_version_min_across_cluster_map_picks_lowest`:
-    /// when every node is at or above the floor, the min-reduce + gate
-    /// composition accepts.
-    #[test]
-    fn parse_valkey_version_all_nodes_at_or_above_floor_accepts() {
-        let body_node1 = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nvalkey_version:8.0.0\r\n";
-        let body_node2 = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nvalkey_version:7.2.0\r\n";
-        let body_node3 = "# Server\r\nredis_version:7.2.4\r\nserver_name:valkey\r\nvalkey_version:9.0.3\r\n";
-        let map = Value::Map(vec![
-            (
-                Value::SimpleString("node1:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node1.to_string(),
-                },
-            ),
-            (
-                Value::SimpleString("node2:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node2.to_string(),
-                },
-            ),
-            (
-                Value::SimpleString("node3:6379".to_string()),
-                Value::VerbatimString {
-                    format: ferriskey::value::VerbatimFormat::Text,
-                    text: body_node3.to_string(),
-                },
-            ),
-        ]);
-
-        let bodies = extract_info_bodies(&map).unwrap();
-        let min = bodies
-            .iter()
-            .map(|b| parse_valkey_version(b).unwrap())
-            .min()
-            .unwrap();
-
-        assert_eq!(min, (7, 2), "min across cluster is the lowest node (7.2)");
-        assert!(
-            min >= (REQUIRED_VALKEY_MAJOR, REQUIRED_VALKEY_MINOR),
-            "all-above-floor cluster must pass the gate"
-        );
-    }
 
     #[test]
     fn valkey_version_too_low_is_not_retryable() {
